Added password change forms and printing views

+++ app/merit/events/utils.py
@@ -5,8 +5,14 @@ from models import RSVP
 # verify_rsvp
 
 def create_rsvp(user, occurrence):
+	if user in occurrence.users.all():
+		return
+
 	rsvp = RSVP(user=user, occurrence=occurrence)
 	rsvp.save()
 
 def destroy_rsvp(rsvp):
+	if not rsvp:
+		return
+
 	rsvp.delete()
